requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
discord.py==2.0.0
fonttools==4.53.1
frozenlist==1.4.1
idna==3.8
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
multidict==6.0.5
nltk==3.9.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pandas_ta==0.3.14b0
pillow==10.4.0
praw==7.7.1
prawcore==2.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
regex==2024.7.24
requests==2.28.1
setuptools==74.1.2
six==1.16.0
ta==0.11.0
textblob==0.18.0.post0
tqdm==4.66.5
tzdata==2024.1
update-checker==0.18.0
urllib3==1.26.20
websocket-client==1.8.0
yarl==1.9.11
